OrderManagement/MicroServices/category/requirements.txt
Django
djangorestframework
django-rest-swagger
djongo
psycopg2
gunicorn
requests
